+++ react/components/src/core/Drawer/DrawerHeader.tsx
@@ -3,8 +3,8 @@ import { createStyles, makeStyles, Theme, useTheme } from '@material-ui/core/sty
 import Toolbar from '@material-ui/core/Toolbar';
 import IconButton from '@material-ui/core/IconButton';
 import Divider from '@material-ui/core/Divider';
-import * as Colors from '@pxblue/colors';
 import { Typography } from '@material-ui/core';
+import PropTypes from 'prop-types';
 
 const useStyles = makeStyles((theme: Theme) =>
     createStyles({
@@ -40,7 +40,7 @@ const useStyles = makeStyles((theme: Theme) =>
         },
         title: {
             fontWeight: 600,
-            lineHeight: '1.25rem',
+            lineHeight: '1.5rem',
         },
         subtitle: {
             fontWeight: 300,
@@ -54,6 +54,11 @@ const useStyles = makeStyles((theme: Theme) =>
             backgroundSize: 'cover',
             height: '100%',
             backgroundPosition: 'center',
+        },
+        nonClickableIcon: {
+            display: 'flex',
+            paddingLeft: theme.spacing(1.5),
+            paddingRight: theme.spacing(1.5),
         },
     })
 );
@@ -76,7 +81,7 @@ export const DrawerHeader: React.FC<DrawerHeaderProps> = (props) => {
     const {
         icon,
         backgroundColor,
-        fontColor,
+        fontColor = theme.palette.getContrastText(backgroundColor || theme.palette.primary.main),
         onIconClick,
         titleContent,
         title,
@@ -85,14 +90,7 @@ export const DrawerHeader: React.FC<DrawerHeaderProps> = (props) => {
         backgroundOpacity,
     } = props;
 
-    // @ts-ignore // TODO: Palette type definition?
-    const toolbarBackgroundColor = String(backgroundColor || theme.palette.primary[500]);
-
-    const clickIcon = (): void => {
-        if (onIconClick) {
-            onIconClick();
-        }
-    };
+    const toolbarBackgroundColor = String(backgroundColor || theme.palette.primary.main);
 
     const getHeaderContent = (): ReactNode =>
         titleContent || (
@@ -129,9 +127,17 @@ export const DrawerHeader: React.FC<DrawerHeaderProps> = (props) => {
                 {getBackgroundImage()}
                 {icon && (
                     <div className={classes.navigation}>
-                        <IconButton color={'inherit'} onClick={(): void => clickIcon()}>
-                            {icon}
-                        </IconButton>
+                        {onIconClick && (
+                            <IconButton
+                                color={'inherit'}
+                                onClick={(): void => {
+                                    onIconClick();
+                                }}
+                            >
+                                {icon}
+                            </IconButton>
+                        )}
+                        {!onIconClick && <div className={classes.nonClickableIcon}>{icon}</div>}
                     </div>
                 )}
                 {getHeaderContent()}
@@ -144,5 +150,15 @@ export const DrawerHeader: React.FC<DrawerHeaderProps> = (props) => {
 DrawerHeader.displayName = 'DrawerHeader';
 DrawerHeader.defaultProps = {
     backgroundOpacity: 0.3,
-    fontColor: Colors.white[50], // TODO: Dark Theme
+};
+DrawerHeader.propTypes = {
+    backgroundColor: PropTypes.string,
+    backgroundImage: PropTypes.string,
+    backgroundOpacity: PropTypes.number,
+    fontColor: PropTypes.string,
+    icon: PropTypes.element,
+    onIconClick: PropTypes.func,
+    subtitle: PropTypes.string,
+    title: PropTypes.string,
+    titleContent: PropTypes.element,
 };